+++ actix-web-lab/src/header.rs
@@ -1,4 +1,5 @@
 //! Experimental typed headers.
 
 pub use crate::cache_control::{CacheControl, CacheDirective};
-pub use crate::hsts::Hsts;
+#[allow(deprecated)]
+pub use crate::strict_transport_security::{Hsts, StrictTransportSecurity};